requirements.txt
flask-cors~=3.0.8
flask~=1.1.2
IMDbPY~=6.8
kinopoiskpy~=0.8.2
